Rework power-ups: tile-based density spawning, stacking buffs, minimap dots
Power-ups now spawn on any empty in-bounds map tile (never on solids/deco
or an already-occupied tile), filling gradually toward a density target of one
per 8x8 tile block, one placed every 6 seconds. Tuning lives in a new
powerup-config.ts (density, cadence, per-type spawn weights, durations).

Timed buffs now stack (re-grabbing adds time, clamped to a uint16 ceiling) and
last much longer: haste 10s, shield 15s, invis 20s, rapidfire/ricochet 30s.
The playerShipTimings packet widens the 5 buff timers from uint8 to uint16 so
durations past 255 ticks survive the wire.

Cloaked ships now fade their health bar and buff rings with the ship sprite, so
a hidden enemy is no longer betrayed. Power-ups also render as colored dots on
the minimap, and the missing rapidfire color was added to the world sprite.

Tile centers follow grid-map.ts geometry (col*cellSize, no half-cell offset).
Tests cover tile placement, no-overlap, density, cadence, stacking, and the
uint16 wire round-trip; regenerated the frozen packet golden and the
deterministic sim digest for the intended wire/spawn changes.

+++ packages/game/src/networking/packets.ts
@@ -166,20 +166,21 @@ export const packetManager = new PacketManager({
         healthRegenerationHeal: $uint8,
         invincibility: $uint8,
         // Timed-buff timers (HASTE_TICKS / SHIELD_TICKS / INVIS_TICKS /
-        // RICOCHET_TICKS / RAPIDFIRE_TICKS). uint8, so durations must stay <= 255
-        // ticks. Networked so remote ships' buffs are known (for the visual + the
-        // tactical powerup feed countdown) and the local player's prediction uses
+        // RICOCHET_TICKS / RAPIDFIRE_TICKS). uint16 (max 65535 ticks) because buff
+        // durations exceed 255 and buffs stack. Networked so remote ships' buffs
+        // are known (for the visual + the tactical powerup feed countdown) and the
+        // local player's prediction uses
         // the same haste. `invisibility` drives the cloak fade and is DISTINCT from
         // the `invincibility` no-damage timer above. `ricochet` is the bouncing-
         // bullet window; the bounce is still resolved server-side on the bullets,
         // but the timer rides the wire so the feed can show its remaining time on
         // enemies. `rapidfire` is the faster-trigger window (the rate scaling is
         // applied locally in PipShip.shoot off this same timer).
-        haste: $uint8,
-        shield: $uint8,
-        invisibility: $uint8,
-        ricochet: $uint8,
-        rapidfire: $uint8,
+        haste: $uint16,
+        shield: $uint16,
+        invisibility: $uint16,
+        ricochet: $uint16,
+        rapidfire: $uint16,
     }),
 
     playerShipCapacities: new Packet({
